+++ src/jllvm/llvm/BitReaderJNI.java
@@ -9,10 +9,10 @@
 package jllvm.llvm;
 
 class BitReaderJNI {
-  public final static native long LLVMParseBitcode(long jarg1, long jarg2, long jarg3);
-  public final static native long LLVMParseBitcodeInContext(long jarg1, long jarg2, long jarg3, long jarg4);
-  public final static native long LLVMGetBitcodeModuleInContext(long jarg1, long jarg2, long jarg3, long jarg4);
-  public final static native long LLVMGetBitcodeModule(long jarg1, long jarg2, long jarg3);
-  public final static native long LLVMGetBitcodeModuleProviderInContext(long jarg1, long jarg2, long jarg3, long jarg4);
-  public final static native long LLVMGetBitcodeModuleProvider(long jarg1, long jarg2, long jarg3);
+  public final static native int LLVMParseBitcode(long jarg1, long jarg2, long jarg3);
+  public final static native int LLVMParseBitcodeInContext(long jarg1, long jarg2, long jarg3, long jarg4);
+  public final static native int LLVMGetBitcodeModuleInContext(long jarg1, long jarg2, long jarg3, long jarg4);
+  public final static native int LLVMGetBitcodeModule(long jarg1, long jarg2, long jarg3);
+  public final static native int LLVMGetBitcodeModuleProviderInContext(long jarg1, long jarg2, long jarg3, long jarg4);
+  public final static native int LLVMGetBitcodeModuleProvider(long jarg1, long jarg2, long jarg3);
 }
